SQLitePlugin: get all integers from sqlite as long ints
Enables storing and retrieving timestamps in the database.

KFA-30, KFA-33

+++ src/android/org/pgsqlite/SQLitePlugin.java
@@ -463,7 +463,7 @@ public class SQLitePlugin extends CordovaPlugin
 									row.put(key, JSONObject.NULL);
 									break;
 								case Cursor.FIELD_TYPE_INTEGER:
-									row.put(key, cur.getInt(i));
+									row.put(key, cur.getLong(i));
 									break;
 								case Cursor.FIELD_TYPE_FLOAT:
 									row.put(key, cur.getFloat(i));
